requirements.txt
flask
flask-cors
requests
groq
python-dotenv
